# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Authentication and Authorization API",
        "title": "Auth API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/pdf/add-page-numbers": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PDF"
                ],
                "summary": "Add page numbers to PDF",
                "parameters": [
                    {
                        "description": "PDF ID and font settings",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AddPageNumbersRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/pdf/add-page-numbers/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PDF"
                ],
                "summary": "Get add page numbers job by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.AddPageNumberJob"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/pdf/compress": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pdf-compress"
                ],
                "summary": "Create compress job",
                "parameters": [
                    {
                        "description": "compress job",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CompressRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/pdf/compress/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pdf-compress"
                ],
                "summary": "Get compress job by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "compress job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CompressJob"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/pdf/crop": {
            "post": {
                "description": "PDF sahifalarini belgilangan tomonlardan qirqadi (top, bottom, left, right)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PDF Edit"
                ],
                "summary": "Crop PDF pages",
                "parameters": [
                    {
                        "description": "Crop request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CropPDFRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/pdf/crop/{id}": {
            "get": {
                "description": "Crop PDF jarayonining natijasini olish",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PDF Edit"
                ],
                "summary": "Get Crop Job info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CropPDFJob"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
        "/api/pdf/watermark": {
            "post": {
                "description": "Add a watermark text to specific pages of a PDF file",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pdf-watermark"
                ],
                "summary": "Add text watermark to PDF",
                "parameters": [
                    {
                        "description": "Watermark request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AddWatermarkRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/pdf/watermark/{id}": {
            "get": {
                "description": "Retrieve watermark job details and status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pdf-watermark"
                ],
                "summary": "Get watermark job by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Watermark Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.AddWatermarkJob"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.AddPageNumberJob": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "first_number": {
                    "type": "integer"
                },
                "font_size": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "input_file_id": {
                    "type": "string"
                },
                "output_file_id": {
                    "type": "string"
                },
                "page_range": {
                    "type": "string"
                },
                "position": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.AddPageNumbersRequest": {
            "type": "object",
            "required": [
                "color",
                "first_number",
                "font_size",
                "input_file_id",
                "page_range",
                "position"
            ],
            "properties": {
                "color": {
                    "type": "string"
                },
                "first_number": {
                    "description": "\u003c-- bu yer o‘zgardi",
                    "type": "integer",
                    "minimum": 1
                },
                "font_size": {
                    "description": "\u003c-- bu yer o‘zgardi",
                    "type": "integer",
                    "minimum": 1
                },
                "input_file_id": {
                    "type": "string"
                },
                "page_range": {
                    "type": "string"
                },
                "position": {
                    "type": "string"
                }
            }
        },
        "models.AddWatermarkJob": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "fill_color": {
                    "type": "string"
                },
                "font_name": {
                    "type": "string"
                },
                "font_size": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "input_file_id": {
                    "type": "string"
                },
                "opacity": {
                    "type": "number"
                },
                "output_file_id": {
                    "type": "string"
                },
                "pages": {
                    "type": "string"
                },
                "position": {
                    "type": "string"
                },
                "rotation": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.AddWatermarkRequest": {
            "type": "object",
            "required": [
                "fill_color",
                "font_name",
                "font_size",
                "input_file_id",
                "position",
                "text"
            ],
            "properties": {
                "fill_color": {
                    "description": "e.g., \"#FF0000\"",
                    "type": "string"
                },
                "font_name": {
                    "description": "e.g., Helvetica",
                    "type": "string"
                },
                "font_size": {
                    "description": "e.g., 24",
                    "type": "integer",
                    "maximum": 72,
                    "minimum": 8
                },
                "input_file_id": {
                    "type": "string"
                },
                "opacity": {
                    "description": "e.g., 0.6",
                    "type": "number",
                    "maximum": 1,
                    "minimum": 0
                },
                "pages": {
                    "description": "optional: \"1-3\"",
                    "type": "string"
                },
                "position": {
                    "description": "e.g., \"tl\"",
                    "type": "string"
                },
                "rotation": {
                    "description": "e.g., 0",
                    "type": "integer",
                    "maximum": 360,
                    "minimum": 0
                },
                "text": {
                    "type": "string"
                }
            }
        },
...
        "models.CompressJob": {
            "type": "object",
            "properties": {
                "compression": {
                    "description": "\"low\", \"medium\", \"high\"",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.CompressionLevel"
                        }
                    ]
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "input_file_id": {
                    "type": "string"
                },
                "output_file_id": {
                    "description": "Output file, can be nil",
                    "type": "string"
                },
                "status": {
                    "description": "\"pending\", \"processing\", \"done\", \"failed\"",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.JobStatus"
                        }
                    ]
                },
                "user_id": {
                    "description": "Can be nil for guest users",
                    "type": "string"
                }
            }
        },
        "models.CompressRequest": {
            "type": "object",
            "required": [
                "compression",
                "input_file_id"
            ],
            "properties": {
                "compression": {
                    "description": "\"low\", \"medium\", \"high\"",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.CompressionLevel"
                        }
                    ]
                },
                "input_file_id": {
                    "type": "string"
                }
            }
        },
        "models.CompressionLevel": {
            "type": "string",
            "enum": [
                "low",
                "medium",
                "high"
            ],
            "x-enum-varnames": [
                "Low",
                "Medium",
                "High"
            ]
        },
...
        "models.CropPDFJob": {
            "type": "object",
            "properties": {
                "bottom": {
                    "description": "Pastki chegara",
                    "type": "integer"
                },
                "box": {
                    "description": "Box turi",
                    "type": "string"
                },
                "created_at": {
                    "description": "Yaratilgan vaqt",
                    "type": "string"
                },
                "id": {
                    "description": "Ish IDsi",
                    "type": "string"
                },
                "input_file_id": {
                    "description": "Kirish fayl IDsi",
                    "type": "string"
                },
                "input_file_ids": {
                    "description": "Fayl IDlari ro'yxati (bir nechta fayl uchun)",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "left": {
                    "description": "Chap chegara",
                    "type": "integer"
                },
                "output_file_id": {
                    "description": "Chiqish fayli IDsi",
                    "type": "string"
                },
                "pages": {
                    "description": "Sahifalar",
                    "type": "string"
                },
                "right": {
                    "description": "O'ng chegara",
                    "type": "integer"
                },
                "status": {
                    "description": "Ish holati (masalan, \"pending\", \"done\")",
                    "type": "string"
                },
                "top": {
                    "description": "Yuqori chegara",
                    "type": "integer"
                },
                "user_id": {
                    "description": "Foydalanuvchi IDsi (guest uchun nil bo'lishi mumkin)",
                    "type": "string"
                }
            }
        },
        "models.CropPDFRequest": {
            "type": "object",
            "required": [
                "input_file_id"
            ],
            "properties": {
                "bottom": {
                    "type": "integer"
                },
                "box": {
                    "description": "optional: mediabox, cropbox, etc",
                    "type": "string"
                },
                "input_file_id": {
                    "type": "string"
                },
                "left": {
                    "type": "integer"
                },
                "pages": {
                    "description": "optional: qaysi sahifalarni crop qilish",
                    "type": "string"
                },
                "right": {
                    "type": "integer"
                },
                "top": {
                    "type": "integer"
                }
            }
        },
...
        "models.JobStatus": {
            "type": "string",
            "enum": [
                "pending",
                "processing",
                "done",
                "failed"
            ],
            "x-enum-varnames": [
                "Pending",
                "Processing",
                "Done",
                "Failed"
            ]
        },
...
        "models.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "description": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}